test: add test to verify that the image changes on hover
EX-5970
